requirements.txt
python-dotenv
zhipuai
pymupdf
requests
flask
flask-cors
